Updated dsui files for using with dstream_filters. Thought I had fixed the bug where the type was lost. We will try again.

+++ branches/EC_DT_RTAS/ace/Counter.h
@@ -16,7 +16,7 @@ class Object_ID
   /* variables */
   uint32_t id;
   uint32_t pid;
-  uint64_t tid;
+  uint32_t tid;
 
   /* EC Specific */
   uint32_t queue_id;
